+++ sql/core/src/main/scala/org/apache/spark/sql/SQLContext.scala
@@ -579,10 +579,9 @@ class SQLContext private[sql](
     DataFrame(self, LocalRelation(attrSeq, rows.toSeq))
   }
 
-
   /**
    * :: Experimental ::
-   * Returns a [[DataFrameReader]] that can be used to read data in as a [[DataFrame]].
+   * Returns a [[DataFrameReader]] that can be used to read data and streams in as a [[DataFrame]].
    * {{{
    *   sqlContext.read.parquet("/path/to/file.parquet")
    *   sqlContext.read.schema(schema).json("/path/to/file.json")
@@ -593,14 +592,6 @@ class SQLContext private[sql](
    */
   @Experimental
   def read: DataFrameReader = new DataFrameReader(this)
-
-
-  /**
-   * :: Experimental ::
-   * Returns a [[DataStreamReader]] than can be used to access data continuously as it arrives.
-   */
-  @Experimental
-  def streamFrom: DataStreamReader = new DataStreamReader(this)
 
   /**
    * :: Experimental ::
